Moved heuristic method computation to state.
